Migrate write tests
